module-1/project-1.1-streamlit/pyproject.toml
[project]
name = "streamlit-tutorial"
version = "0.1.0"
description = "Streamlit tutorial with factorial calculator and grade analysis"
authors = [
    {name = "Dang Le", email = "lehaidang2601@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "openpyxl>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
